requirements.txt
fastapi
uvicorn[standard]
python-multipart
httpx
chromadb
pydantic
torch
transformers
pillow
sentence-transformers
openai
SPARQLWrapper